+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Log portable actor registration errors.
 
+### Fixed
+
+- Fix registration of portable actors.
+
 ## [3.7.0] - 2024-10-16
 
 ### Added